+++ src/error.rs
@@ -29,6 +29,18 @@ pub enum Error {
     /// conversion failed.
     // FIXME: Add the inner Utf8Error here?
     InvalidUtf8InBody,
+    /// Tried to send a secure request (ie. the url started with
+    /// `https://`), but the crate's `https` feature was not enabled,
+    /// and as such, a connection cannot be made.
+    HttpsFeatureNotEnabled,
+    /// The provided url contained a domain that has non-ASCII
+    /// characters, but it could not be converted into punycode
+    /// because the `punycode` feature was not enabled.
+    PunycodeFeatureNotEnabled,
+    /// The provided url contained a domain that has non-ASCII
+    /// characters, and could not be converted into punycode. It is
+    /// probably not an actual domain.
+    PunycodeConversionFailed,
 
     /// This is a special error case, one that should never be
     /// returned! Think of this as a cleaner alternative to calling
@@ -49,6 +61,9 @@ impl fmt::Display for Error {
             InfiniteRedirectionLoop => write!(f, "infinite redirection loop detected"),
             InvalidUtf8InResponse => write!(f, "response contained invalid utf-8 where valid utf-8 was expected"),
             InvalidUtf8InBody => write!(f, "response body contains invalid utf-8, so it can't be converted into a string"),
+            HttpsFeatureNotEnabled => write!(f, "request url contains https:// but the https feature is not enabled"),
+            PunycodeFeatureNotEnabled => write!(f, "non-ascii urls needs to be converted into punycode, and the feature is missing"),
+            PunycodeConversionFailed => write!(f, "non-ascii url conversion to punycode failed"),
             Other(msg) => write!(f, "error in minreq: please open an issue in the minreq repo, include the following: '{}'", msg),
 
             #[cfg(feature = "json-using-serde")]